refactor: refactored wallet-web folder structure
