+++ cmd/wallet-web/src/pages/NotFound.vue
