+++ cmd/wallet-web/src/pages/PresentationDefQuery.vue
@@ -300,7 +300,7 @@ import {
   getCredentialDisplayData,
   getCrendentialIcon,
   WalletGetByQuery,
-} from './mixins';
+} from '@/utils/mixins';
 import { mapGetters } from 'vuex';
 import Spinner from '@/components/Spinner/Spinner.vue';
 import { useI18n } from 'vue-i18n';
