--- cmd/wallet-web/src/pages/layout/DashboardContent.vue
@@ -1,15 +1,0 @@
-<!--
- * Copyright SecureKey Technologies Inc. All Rights Reserved.
- *
- * SPDX-License-Identifier: Apache-2.0
--->
-
-<template>
-  <router-view></router-view>
-</template>
-
-<script>
-export default {
-  name: 'DashboardContent',
-};
-</script>
