+++ cmd/wallet-web/src/router/routes.js
@@ -5,69 +5,45 @@ SPDX-License-Identifier: Apache-2.0
 */
 
 // Lazy load the component
-function load(component) {
-  return () => import(/* webpackChunkName: "[request]" */ `@/pages/${component}.vue`);
+function load(path) {
+  return () => import(`@/${path}.vue`);
 }
 
 export default [
   {
     path: '',
     name: 'DashboardLayout',
-    component: load('layout/DashboardLayout'),
+    component: load('layouts/DashboardLayout'),
     redirect: 'vaults',
     children: [
       {
         path: 'vaults',
         name: 'vaults',
-        component: load('Vaults'),
+        component: load('pages/Vaults'),
         meta: { requiresAuth: true },
       },
       {
         path: 'credentials',
         name: 'credentials',
-        component: load('Credentials'),
+        component: load('pages/Credentials'),
         meta: { requiresAuth: true },
       },
       {
         path: 'credentials/:id',
         name: 'credential-details',
-        component: load('CredentialDetails'),
+        component: load('pages/CredentialDetails'),
         meta: { requiresAuth: true },
       },
       {
         path: 'waci',
         name: 'waci',
-        component: load('WACI'),
+        component: load('layouts/WACI'),
         meta: { requiresAuth: true, signin: true, disableCHAPI: true },
       },
       {
         path: 'did-management',
         name: 'did-management',
-        component: load('DIDManagement'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'connections',
-        name: 'connections',
-        component: load('demos/Connections'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'relationships',
-        name: 'relationships',
-        component: load('demos/Relationships'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'issue-credential',
-        name: 'issue-credential',
-        component: load('demos/IssueCredential'),
-        meta: { requiresAuth: true },
-      },
-      {
-        path: 'present-proof',
-        name: 'present-proof',
-        component: load('demos/PresentProof'),
+        component: load('pages/DIDManagement'),
         meta: { requiresAuth: true },
       },
     ],
@@ -75,44 +51,44 @@ export default [
   {
     path: 'StoreInWallet',
     name: 'chapi-store',
-    component: load('Store'),
+    component: load('pages/Store'),
     meta: { blockNoAuth: true },
   },
   {
     path: 'GetFromWallet',
     name: 'chapi-get',
-    component: load('Get'),
+    component: load('layouts/Get'),
     meta: { blockNoAuth: true, isNavbarHidden: true },
   },
   {
     path: 'worker',
     name: 'chapi-worker',
-    component: load('Worker'),
+    component: load('pages/Worker'),
   },
   {
     path: 'loginhandle',
     name: 'loginhandle',
-    component: load('LoginHandle'),
+    component: load('pages/LoginHandle'),
     props: (route) => ({ provider: route.query.provider }),
   },
   {
     path: 'signin',
     name: 'signin',
-    component: load('Signin'),
+    component: load('pages/Signin'),
   },
   {
     path: 'signup',
     name: 'signup',
-    component: load('Signup'),
+    component: load('pages/Signup'),
   },
   {
     path: 'needauth',
     name: 'block-no-auth',
-    component: load('BlockNoAuth'),
+    component: load('pages/BlockNoAuth'),
   },
   {
     path: 'pathMatch(.*)*',
-    name: 'PageNotFound',
-    component: load('PageNotFound'),
+    name: 'NotFound',
+    component: load('pages/NotFound'),
   },
 ];
