+++ cmd/wallet-web/src/utils/mixins/common/helper.js
