+++ cmd/wallet-web/src/utils/mixins/common/register.js
