+++ cmd/wallet-web/src/utils/mixins/get/didConn.js
