+++ cmd/wallet-web/src/utils/mixins/get/getCredentialsByQuery.js
