+++ cmd/wallet-web/src/utils/mixins/index.js
