+++ cmd/wallet-web/test/specs/store.spec.js
@@ -6,10 +6,10 @@ SPDX-License-Identifier: Apache-2.0
 
 import { nextTick } from 'vue';
 import { shallowMount } from '@vue/test-utils';
-import Store from '../../src/pages/Store.vue';
+import Store from '@/pages/Store.vue';
 import { getTestData, Setup, MockCredentialHandler, promiseWhen } from '../common';
 import { expect } from 'chai';
-import i18n from '../../src/plugins/i18n';
+import i18n from '@/plugins/i18n';
 
 const SAVE_CRED_USER = 'save_cred_user';
